chore: move console script into its own file
